refactor: 💡 file names, as export names and some chore
